Refactor App component and add Checkbox component
This commit refactors the App component in src/App.jsx to use the new Checkbox component. The Checkbox component is imported from src/components/Checkbox.jsx and is used to render a list of checkboxes based on the data in the "list" array. Each checkbox displays a text and a completion status based on the "isChecked" property. The Checkbox component also includes PropTypes for type checking.